PQPW release baseline: harden native Linux renderer

+++ World/World.h
@@ -30,24 +30,24 @@
 #ifndef GE_WORLD_H
 #define GE_WORLD_H
 
-#include "ErrorLog.h"
-#include "PtrTypes.h"
+#include "Errorlog.h"
+#include "Ptrtypes.h"
 #include "Genesis.h"
-#include "GBSPFile.h"
-#include "Motion.h"
-#include "Surface.h"
+#include "Gbspfile.h"
+#include "motion.h"
+#include "SURFACE.H"
 #include "Fog.h"
 #include "WBitmap.h"
-#include "User.h"
+#include "USER.H"
 #include "Light.h"
 
-#include "Bitmaplist.h"
+#include "BitmapList.h"
 
-#include "Actor.h"			
+#include "actor.h"
 
 //MRB BEGIN
 //geSprite
-#include "Sprite.h"
+#include "sprite.h"
 //MRB END			
 
 #ifdef __cplusplus
@@ -318,7 +318,7 @@ geBoolean geWorld_DetachAll(geWorld *World);
 
 /* 03/24/2004 Wendell Buckner
     BUG FIX: Rendering Transparent Polys properly (2) */
-void GList_AddOperation(uint8 Type, uint32 Data);
+void GList_AddOperation(uint8 Type, uintptr_t Data);
 
 /* 02/18/2004 Wendell Buckner
     DOT BUMPMAPPING */
@@ -327,10 +327,9 @@ GENESISAPI geBoolean geWorld_CreateBumpmapByNameDot3(geWorld *World, const char 
 /* 10/15/2003 Wendell Buckner
     Bumpmapping for the World */
 GENESISAPI geBitmap * geWorld_CreateBumpmapByName(geWorld *World, const char *BumpmapName, gePixelFormat BumpFormat);
-GENESISAPI geBoolean geWorld_GetWorldGeometry(geWorld *World, geVec3d **Verts, int *NumVerts, long **Indices, long *NumIndices);
+GENESISAPI geBoolean geWorld_GetWorldGeometry(geWorld *World, geVec3d **Verts, int32 *NumVerts, int32 **Indices, int32 *NumIndices);
 
 #ifdef __cplusplus
 }
 #endif
 #endif
-
